+++ docs/changelog.rst
@@ -4,6 +4,9 @@ Changelog
 1.0b15 (in development)
 -----------------------
 
+* The default output format for the ``beet list`` command is now configurable
+  via the :ref:`list_format_item <list_format_item>` and :ref:`list_format_album
+  <list_format_album>` config options. Thanks to Fabrice Laporte.
 * New plugin event: ``import_task_choice`` is called after an import task has an
   action assigned.
 * New plugin event: ``library_opened`` is called when beets starts up and